Fix circular dependancy in addPages
